- file will exist only on error (thx tony for the headup)
git-svn-id: http://svn.php.net/repository/php/php-src/php/php-src/trunk@265693 c90b9560-bf6c-de11-be94-00142212c4b1

+++ ext/zip/tests/oo_delete.phpt
@@ -63,8 +63,11 @@ $sb = $zip->statIndex(1);
 var_dump($sb);
 $sb = $zip->statIndex(2);
 var_dump($sb);
+$zip->close();
 unset($zip);
-unlink($file);
+if (file_exists($file)) {
+	unlink($file);
+}
 ?>
 --EXPECTF--
 ok
